requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
pdfplumber==0.8.0
PyPDF2==3.0.1
spacy==3.5.1
defusedxml~=0.7.1
cffi~=1.15.1
Pillow~=9.4.0
ipython~=8.11.0
jupyter~=1.0.0
pip~=22.3.1
attrs~=22.2.0
distro~=1.8.0
wheel~=0.38.4
cryptography~=40.0.1
tornado~=6.2
Jinja2~=3.1.2
pyzmq~=25.0.2
pexpect~=4.8.0
blis~=0.7.9
comm~=0.1.3
traitlets~=5.9.0
parso~=0.8.3
jedi~=0.18.2
pytz~=2023.2
Pygments~=2.14.0
requests~=2.28.2
ipywidgets~=8.0.6
notebook~=6.5.3
Wand~=0.6.11
cymem~=2.0.7
setuptools~=65.5.1
pathy~=0.10.1
typer~=0.7.0
thinc~=8.1.9
srsly~=2.4.6
wasabi~=1.1.1
click~=8.1.3
tqdm~=4.65.0
catalogue~=2.0.8
pydantic~=1.10.7
packaging~=23.0
langcodes~=3.3.0
preshed~=3.0.8
psutil~=5.9.4
murmurhash~=1.0.9
confection~=0.0.4
jsonschema~=4.17.3
PyYAML~=6.0
MarkupSafe~=2.1.2
python-dateutil~=2.8.2
certifi~=2022.12.7
debugpy~=1.6.6
decorator~=5.1.1
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.6
ptyprocess~=0.7.0
cycler~=0.11.0
six~=1.16.0
idna~=3.4
urllib3~=1.26.15
asttokens~=2.2.1
zipp~=3.15.0
contourpy~=1.0.7
executing~=1.2.0
fonttools~=4.39.3
ipykernel~=6.22.0
pycparser~=2.21
pyparsing~=3.0.9
pyrsistent~=0.19.3
webcolors~=1.13
jsonpointer~=2.3
isoduration~=20.11.0
fqdn~=1.5.1
langdetect~=1.0.9
kiwisolver~=1.4.4
platformdirs~=3.2.0
regex~=2023.3.23
seaborn~=0.12.2
tabula-py~=2.7.0